requirements.txt
langchain
langgraph
langgraph-checkpoint-mongodb
pymongo[asyncio]
langchain-community
langchain-huggingface
langchain-groq
langchain-tavily
python-dotenv
pydantic
einops
unstructured[all-docs]
langchain-unstructured
fastapi
uvicorn[standard]
qdrant-client
fastembed
langchain-qdrant
langchain-deepinfra
python-docx
openpyxl
python-pptx
pyyaml 